NAS-129425: Fix remarks
