+++ src/app/pages/dashboard/types/widget-component.interface.ts
@@ -1,4 +1,5 @@
 import { InputSignal, Type } from '@angular/core';
+import { HiddenWidgetsService } from 'app/pages/dashboard/services/hidden-widgets.service';
 import { WidgetCategory } from 'app/pages/dashboard/types/widget-category.enum';
 import { WidgetSettingsRef } from 'app/pages/dashboard/types/widget-settings-ref.interface';
 import {
@@ -37,6 +38,7 @@ export interface WidgetDefinition<
   category: WidgetCategory;
   component: Type<Component>;
   settingsComponent: SettingsComponent extends null ? null : Type<SettingsComponent>;
+  hidden?: (hiddenService: HiddenWidgetsService) => boolean;
 }
 
 /**
@@ -48,6 +50,7 @@ export interface WidgetDefinition<
  *   category: WidgetCategory.Network,
  *   settingsComponent: WidgetInterfaceIpSettingsComponent,
  *   supportedSizes: [SlotSize.Full, SlotSize.Half, SlotSize.Quarter],
+ *   hidden: (hiddenService) => hiddenService.isHidden(WidgetType.Ipv4Address),
  * });
  * ```
  */
