+++ src/app/pages/dashboard/widgets/system/widget-sys-info-passive/widget-sys-info-passive.definition.ts
@@ -1,7 +1,7 @@
 import { marker as T } from '@biesbjerg/ngx-translate-extract-marker';
 import { WidgetCategory } from 'app/pages/dashboard/types/widget-category.enum';
 import { dashboardWidget } from 'app/pages/dashboard/types/widget-component.interface';
-import { SlotSize } from 'app/pages/dashboard/types/widget.interface';
+import { SlotSize, WidgetType } from 'app/pages/dashboard/types/widget.interface';
 import { WidgetSysInfoPassiveComponent } from 'app/pages/dashboard/widgets/system/widget-sys-info-passive/widget-sys-info-passive.component';
 
 export const systemInfoPassiveWidget = dashboardWidget({
@@ -10,4 +10,5 @@ export const systemInfoPassiveWidget = dashboardWidget({
   category: WidgetCategory.SystemInfo,
   component: WidgetSysInfoPassiveComponent,
   settingsComponent: null,
+  hidden: (hiddenService) => hiddenService.isHidden(WidgetType.SystemInfoPassive),
 });
